# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Stock Analysis Platform API
  version: 1.0.0
  description: RESTful API for stock analysis and portfolio management
  contact:
    name: API Support
    email: support@example.com
servers:
  -
    url: "http://localhost:3000"
    description: Development server
  -
    url: "https://api.example.com"
    description: Production server
paths:
...
  /api/sector/:sector:
    get:
      summary: Get news for a specific sector
      description: "Path Parameters: - sector: Sector name (required) Query Parameters: - page: Page number (optional, default: 1) - limit: Items per page (optional, default: 20, max: 100)"
      tags:
        - News
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: sector
          in: path
          description: Sector name
          required: true
          schema:
            type: string
        -
          name: page
          in: path
          description: Page number
          required: false
          schema:
            type: string
        -
          name: limit
          in: path
          description: Items per page
          required: false
          schema:
            type: string
...
  /api/subscriptions:
    get:
      summary: 获取用户订阅的板块列表
      description: null
      tags:
        - Sector
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
...
  /api/:sectorId/subscribe:
    post:
      summary: 订阅板块
      description: null
      tags:
        - Sector
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                      id: 1
                      createdAt: "2024-01-01T00:00:00Z"
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: sectorId
          in: path
          description: sectorId parameter
          required: true
          schema:
            type: string
    delete:
      summary: 取消订阅板块
      description: null
      tags:
        - Sector
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                      deleted: true
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: sectorId
          in: path
          description: sectorId parameter
          required: true
          schema:
            type: string
  /api/:sectorId/stocks:
    get:
      summary: 获取板块内的股票列表
      description: null
      tags:
        - Sector
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                      symbol: AAPL
                      name: Apple Inc.
                      price: 150.25
                      change: 2.5
                      volume: 1000000
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: sectorId
          in: path
          description: sectorId parameter
          required: true
          schema:
            type: string
  /api/:sectorId/news:
    get:
      summary: 获取板块相关新闻
      description: null
      tags:
        - Sector
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: sectorId
          in: path
          description: sectorId parameter
          required: true
          schema:
            type: string
  /api/:sectorId/performance:
    get:
      summary: 获取板块表现数据
      description: null
      tags:
        - Sector
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: sectorId
          in: path
          description: sectorId parameter
          required: true
          schema:
            type: string
...
  /api/market/sectors:
    get:
      summary: Get list of available sectors for heatmap filtering
      description: "IMPORTANT: This route MUST be defined BEFORE /:symbol routes"
      tags:
        - Stocks
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...